meta-crownbay: add a machine configuration that doesn't use EMGD graphics
This adds a second machine to the meta-crownbay BSP that uses generic
vesa graphics instead of the Intel-proprietary straight 'crownbay'
version that does.

The name of the new machine is 'crownbay-noemgd'.

Though the graphics quality isn't as nice, the binaries for this
version can be distributed without any need for license click-throughs
or the manual step currently required for the EMGD version.

Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>