n450: revamp audio mixer init script
Fixes [YOCTO #798]

n450-audiofix suffered from various problems, including not getting installed.
This patch adds it via MACHINE_EXTRA_RRECOMMENDS and cleans up the various
bugs in the implementation. The structure has been renamed and the comments
updated for consistency. This should be considered a stop-gap measure, as
something along the lines of oe's alsa-state is a far better solution.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>