linux-yocto*: correct FILESEXTRAPATHS assignment in all layers
My previous commit used a self referencing immediate expansion assignment using.
This causes a circular reference and build failure, particularly when more than
one layer is configured in.

Use the _prepend mechanism in conjunction with the immediate expansion
assignment operator to accomplish a similar thing without the self referencing.

Note: This patch is dependent on changes to oe-core, specifically:
http://lists.linuxtogo.org/pipermail/openembedded-core/2011-May/003145.html

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Cc: Tom Zanussi <tom.zanussi@intel.com>
Cc: Phill Blundell <pb@pbcl.net>
Cc: Koen Kooi <koen@dominion.thruhere.net>