meta-intel/linux-yocto: update to match the renamed linux-yocto recipes
The linux-yocto recipes have been renamed to include explicit
version numbers. The meta-intel bbappends must be updated to
follow the new naming convention.

As a consequence of the recipe renames any SRCREV specificiations
that were part of a machine.conf have been moved to the corresponding
bbappend file. This allows the 2.6.34 and 2.6.37 SRCREVs to co-exist.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>

Minor whitespace cleanup.
Remove inadvertant? linux-libc-headers provider change.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>