meta-intel/bernard: Fix compile_perf build breakage in linux-yocto-stable for crownbay.
This way new users of yocto can start with a working default configuration for crownbay.

The issue is described in more detail in the following thread  on the
yocto mailing list.

http://lists.yoctoproject.org/pipermail/yocto/2011-July/002129.html

[tom.zanussi@intel.com] The crownbay branch in the repo missed getting
the standard branch containing the NO_LIBPERL/NO_LIBPYTHON merged in,
which it now has been, so we need to update the SRCREVs and pick it
up.

Signed-off-by: Andre Haupt <andre@bitwigglers.org>
Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>