meta-tlk: initial creation
Call this layer whatever you want, but when enabled it turns on BSP configuration
items specific to autobuild/official binary images. Once this is enabled the uptime
of any BSP will be limted to the value in the .cfg. It is currently at 2 minutes
as an example, but it should be extended to a number of days going forward.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>

meta-tlk: Rename and changed to 10 day limit

Signed-off-by: Saul Wold <sgw@linux.intel.com>