rt: simplify linux-yocto-rt.bbappend for all BSPs
As the all the BSPs use the same BSP branch and meta commit ID as the base
recipe, there is no need specify the KBRANCH and SRCREVs. Not doing so
greatly simplifies maintenance. Leaving the syntax in place, but commented
out, serves as documentation should the need arise to use a different BSP
branch or meta commit ID.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>