xpsb-glx: Move the pkgconfig file(s) to the -dev package
pkgconfig files belong in the -dev package.  They automatically add
a runtime dependency (when packaging with RPM) for "pkgconfig", which
is not desired to be installed with this package.

This also fixes another problem in that FILES_{PN} was being assigned
instead of being appended and was adding libdir/*, preventing the
above from taking effect.

Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>