libva: upgrade to 1.0.12 and factor out common settings
EMGD 1.8 requires libva-1.0.12, so upgrade to that version, and we
don't actually need the 1.0.10 recipe, so remove it.

This also adds -dbg and -dev packaging fixes to fix QA problems,
removes some unnecessary settings, updates checksums, etc.

Also create and use a common libva.inc, as we'll soon be adding a more
updated version of libva to the current version.

Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>