meta-emenlow: fix ptr->int and Werror=address compile errors
A couple of things that had previously been warnings are now errors,
so they need to be fixed up.

The first problem is the same problem encountered by xserver-kdrive,
and has the same fix, namely create a new version of REGION_INIT and
have all calls that pass in a pointer to a static struct use it
instead.  A different fix for this is upstream, which includes an API
change; we don't need all that to fix this problem.

The second problem is a cast from pointer to integer in fbdevhw.c.
This also is fixed upstream by removing the whole section of code
which is bogus anyway, which is also done here.

Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>