fri2: Autoload module iwlwifi
iwlwifi uses a firmware blob that must exist on the FS at the time
the driver is loaded. If the driver is built-in, the blob must exist
in the initrd. We avoid this by building as a module, but we still
want this to load at boot - so add it to autoload.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>