efilinux: Add efilinux 1.0 and gnu-efi
efilinux is a UEFI OS loader. It was created as a reference implementation with
the aim of being well documented and containing well written source code.

efilinux is lightweight and convenient as a tool to debug misbehaving UEFI
systems.

Thanks to Paul Eggleton for is initial set of recipes from which these were
based.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
CC: Paul Eggleton <paul.eggleton@linux.intel.com>
CC: Tom Zanussi <tom.zanussi@intel.com>
CC: Kishore K Bodke <kishore.k.bodke@intel.com>
Signed-off-by: Darren Hart <dvhart@linux.intel.com>