meta-sugarbay: use rc6 kernel feature
According to the Intel graphics guys and Ubuntu, it's ok to now have
rc6 enabled by default (as long as rc6p is not also enabled).

Accordingly, we'll use the rc6 feature that enables it.  If there are
problems, it can be easily disabled at build-time by removing tmp/rc6
from KERNEL_FEATURES, or can be disbled at run-time using the kernel
parameter i915_enable_rc6 = 0.

Signed-off-by: Tom Zanussi <tom.zanussi@intel.com>