nuc: Update linux-yocto 3.10 SRCREV (3.10.34-LTSI)
Update to the HEAD of standard/base (3.10.34-LTSI). This also works
around an open issue with do_validate_branches where feature branches
are reset the HEAD of the machine branch if they contain that commit.

The common linux-yocto_3.10 kernel is not able to find root device on
the NUC platform.  Hence the NUC BSP is not yet updated to use the
common kernel.

Signed-off-by: Nitin A Kamble <nitin.a.kamble@intel.com>
Acked-by: Darren Hart <dvhart@linux.intel.com>