meta-valleyisland: Add linux-yocto 3.10 recipe
Added linux-yocto 3.10 due to it cannot be shared with
common intel recipe kernel. The reason cannot be shared as for
valleyisland case has additional feature that is still pending to
pull into LTS/LTSI and hence it cannot be merged into "standard/base".
For this particular reason, a feature branch of "valleyisland-io-1.0"
is introduced to include additional feature. However in order to make
it more align with intel common recipe kernel,machine branch is
pointing to "standard/base" and SRCREV for meta remains closest as
possible.

Signed-off-by: Chan Wei Sern <wei.sern.chan@intel.com>
Signed-off-by: Chang Rebecca Swee Fun <rebecca.swee.fun.chang@intel.com>
Acked-By: Nitin A Kamble <nitin.a.kamble@intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>