meta-crystalforest: Intel QAT v1.5 recipe under recipes-extended
This is an Initial version of Intel QuickAssist Technology(QAT)
recipe support. This recipe is targeting on Intel QAT v1.5.

Intel's QuickAssist Technology is designed to optimize the use
and deployment of algorithm accelerators in networking and security applications.

"QAT1.5" is for use with Intel Communications Chipset 8900 to 8920(DH89xxCC)Series.
Therefore, this recipe supports QAT on Cave Creek platforms.
This recipe is created under meta-intel/meta-crystalforest.

Signed-off-by: venkatasubramanian pattu <venkatasubramanianx.pattu@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>