intel-microcode: bump version to 20150121
Bumped version to 20150121. Dropped runtime dependency on iucode-tool
which is used to generate the cpio or trim down the data file based on
your CPU. While it can also be used to load the microcode that is not
the recommended method for Linux 3.9 and newer instead you should use
the kernel's built-in early microcode loading facilities.

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>