gstreamer-vaapi: Update to upstream version 1.8.2
With the 1.8.x series gstreamer-vaapi got integrated into the
maintainership of the greater GStreamer project.  Also this
upstreaming introduced a policy change in the source base where
gstreamer-vaapi versions are now tested only with the GStreamer
version with the same number (as opposed to the past pre-upstream
source base which had a policy of supporting a range of GStreamer
versions).  So care should be taken to keep the gstreamer-vaapi
build configuration in sync with the gstreamer build configuration.

Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>