meta-isg: openssl-qat: fix for patch error
openssl-qat is different from other packages in the way that it
has a patch present in a zip file that is unpacked and applied
on top of openssl. This results in patch errors when doing an
incremental build using bitbake.

To resolve, splitted do_patch into two - one to unpack the qat patch
and apply it on openssl. And, other to apply the patches added in
SRC_URI.

Signed-off-by: Rahul Kumar Gupta <rahul.kumarxx.gupta@intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>