rmc: Add Runtime Machine Configuration (RMC) project
RMC recipe fetch RMC project and build it more than once in
build time:

RMC tool is built for host architecture (native). The tool for
host is used to generate RMC database in build time.

RMC tool is also built for target architecture, so that scripts
in user space can call RMC tool on a running target. Developers
can also boot a target and run rmc tool to obtain fingerprint
for a new board type.

RMC libraries are compiled for both of UEFI context and user
space. They are always linked in RMC tool and can be linked
into an EFI bootloader. The recipes don't install libraries
for target's user space until we have a new client needs it.

The rmc-db.bbclass provides functions to generate rmc database
file for other software components to reuse.

We absorb a patch from Tom Zanussi to update source location with
the public link. We could put this change in another commit, but
leaving the replaced internal link in this commit could cause
trouble when people bisect the project but don't have access to
the internal location:
----------------------
 rmc: Update to use public repo

 The repo the rmc recipe was pointing to was private - it's now public
 Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>

 common/recipes-bsp/rmc/rmc.inc | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

 diff --git a/common/recipes-bsp/rmc/rmc.inc b/common/recipes-bsp/rmc/rmc.inc
 index c046e2e..bdf930d 100644
 --- a/common/recipes-bsp/rmc/rmc.inc
 +++ b/common/recipes-bsp/rmc/rmc.inc
 @@ -15,7 +15,7 @@ LICENSE = "MIT"

 LIC_FILES_CHKSUM = "file://COPYING;md5=bcdd376d27b26bde6afadd67aa3c8b07"

 -SRC_URI = "git://git@git.yoctoproject.org/rmc;protocol=ssh"
 +SRC_URI = "git://git.yoctoproject.org/rmc"
...
----------------------

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>
Reviewed-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>