rmc: Explicitly specify hash style for linker
The QA sanity checker in OE fails when an ELF binary has a symbol
table without GNU hash.

There is no issue so far because rmc is statically linked with C
libraries and the symbol table is stripped out in build.

But once we compile rmc with dynamic linking, the symbol table in the
dynamic section stays. Thus it fails the QA checker because the
default hash style is sysv.

To have some peace of mind, we specify "both" for the hash style to
the linker.

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>