intel-core*: Add support for wic images
This adds WIC Image support to the core2 and corei7 Intel BSPs, these
are all designed to be copied to USB sticks via dd.  This will create
an efi boot partition along with a user space partition.

It is not currently setup to have an install target, it only has a live boot
target.

Signed-off-by: Saul Wold <sgw@linux.intel.com>