rmc: Add board data for NUC5i5RYB
A particular intention of this change is to remove the serial
setting in kernel command line that is hardcoded in core2 and
corei7 conf files.

However we cannot use KBOOTPARAM because it is designed as a
fragment appended to hardcoded part at runtime. This is why
we need to create more files in this patch to use cmdline
in bootloader entry files to bypass whatever from the build.

Once we have a clean base (no any preset console setting)
in generic configurations, this change can be dropped.

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>