rmc: Add example for Galileo Gen 2 (Quark)
This is an experimental RMC sample for quark. At runtime, a
boot option with board name will be available for user to
boot the Galileo.

Note: The supported image type is hddimg instead of the direct-
boot image type usually used on quark. Use a USB stick flashed
with the image to boot the Galileo Gen 2 board.

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>