intel-*: build dosfs/mtools for wic image creation
intel-* machines default to an EFI bootloader which runs from an EFI
system partition (ESP). Also, the ESP needs to be vfat formatted.

Make sure the dosfs/mtools used by wic and its plugins that deal with
vfat are built and available (without depending on do_bootimg) when
wic images are being created.

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>
Signed-off-by: California Sullivan <california.l.sullivan@intel.com>