cryptodev-module: Fix the check
A case was found where the PREFERRED_VERSION_linux-intel was not set
even with linix-intel being the PREFERRED_PROVIDER and this code fails.
So, just check for linux-intel now, we know until the next LTS version
comes out that this kernel will have the API change patch.  A more detailed
check may need to be added.

Signed-off-by: Saul Wold <sgw@linux.intel.com>