conf/machine/intel-core*: Fix WKS_FILE setting
bb.utils.contains checks if ALL items are in the variable, and since we
check for systemd-boot or rmc-boot, it would always be false, assigning
mkefidisk as the WKS_FILE in all cases. bb.utils.contains_any checks if
one or more items are in the variable, so use that instead.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>