iwlwifi_git.bb: run make_scripts earlier
This out-of-tree module requires we create scripts before configuration
but module.bbclass runs make scripts sometime before do_compile. In
some builds this results in a failure, as it might not be soon enough.

Run make scripts again but earlier. Ideally we change module.bbclass
instead, but its currently frozen for M2 testing, so we need this
temporary local fix.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>