recipes-selftest: add images for use in secureboot selftests
This patch adds two image recipes and several keys for use in secureboot
selftests. One image is an unsigned comboapp with a startup.nsh file
calling bootx64.efi. The other is a comboapp image signed with the
refkit keys by default. These are the same keys enrolled in the ovmf
firmware. Also included is another set of keys to sign the image with
that do not match the ovmf firmware's enrolled keys.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>