linux-firware_%.bbappend: multiple fixes
Fix yocto project compatibility issue: was using raw _appends, without a
machineoverride.

Fix not pinning to a specific commit ID. The firmware could be updated,
which would change the hash and break the build temporarily.

Fix naming. There are several similarly named firmware blobs (e.g.,
iwlwifi-8265-31) so not using the real name of the blob was confusing.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>