conf: override WKS_FILE_DEPENDS for intel machines
WKS_FILE_DEPENDS includes all the dependencies for producing wic images
and is meant to be overridden with correct set by image recipes. Right now,
the default values result in grub-efi being built even when EFI_PROVIDER
is set to systemd-boot.

Change the value to depend only on the EFI_PROVIDER bootloader.

Signed-off-by: Anuj Mittal <anuj.mittal@intel.com>
Signed-off-by: California Sullivan <california.l.sullivan@intel.com>