clang: build Intel common-clang and spirv
Common clang is a thin wrapper library around clang. Common clang has
OpenCL-oriented API and is capable to compile OpenCL C kernels to SPIR-V
modules.

This adds a bbappend to clang recipe from meta-clang to build the
necessary components and moves it to dynamic layers so it's built only
when clang-layer is included.

Signed-off-by: Anuj Mittal <anuj.mittal@intel.com>