ispc: add recipe
Intel(R) Implicit SPMD Program Compiler (Intel(R) ISPC)

ispc is a compiler for a variant of the C programming language, with
extensions for single program, multiple data programming. Under the SPMD
model, the programmer writes a program that generally appears to be a
regular serial program, though the execution model is actually that a
number of program instances execute in parallel on the hardware.

https://ispc.github.io/downloads.html
https://github.com/ispc/ispc

Signed-off-by: Naveen Saini <naveen.kumar.saini@intel.com>
Signed-off-by: Anuj Mittal <anuj.mittal@intel.com>