qemu-intel.inc: backport of dependency fixes from oe-core
When building an image for intel-corei7-64 (and probably other targets)
it was not possible to directly start the image with runqemu due to
missing directory in sysroot related to qemu-helper-native.

In oe-core two patches has been applied that fixes dependency issues when
building images for qemu. Those patches does also fix the issue when
building for targets in meta-intel.

Following two patches from oe-core, originally for meta/conf/machine/include/qemu.inc,
are backported to conf/machine/include/qemu-intel.inc:

3a4fed4ae0 qemu.inc: Should depend on qemu-system-native, not qemu-native
5562342020 image/qemu: Add explict depends for qemu-helper addto_recipe_sysroot task

Signed-off-by: Peter Bergin <peter@berginkonsult.se>
Signed-off-by: Anuj Mittal <anuj.mittal@intel.com>